python/projectDjango/NewsPaper/requirements.txt
Django>=3.2,<3.3
django-allauth
django-filter
